Added TravelFundTC. refactored the packageName.
