--- src/main/goldManS/package-info.java
@@ -1,8 +1,0 @@
-/**
- * 
- */
-/**
- * @author Yashkumar.Thakur
- *
- */
-package goldManS;
