+++ src/main/gold_man_s/ArrayMinusOp.java
